dom/webauthn/tests/test_webauthn_loopback.html

<!DOCTYPE html>
<meta charset=utf-8>
<head>
  <title>Full-run test for MakeCredential/GetAssertion for W3C Web Authentication</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="u2futil.js"></script>
  <script type="text/javascript" src="pkijs/common.js"></script>
  <script type="text/javascript" src="pkijs/asn1.js"></script>
  <script type="text/javascript" src="pkijs/x509_schema.js"></script>
  <script type="text/javascript" src="pkijs/x509_simpl.js"></script>
  <script type="text/javascript" src="cbor/cbor.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>

<h1>Full-run test for MakeCredential/GetAssertion for W3C Web Authentication</h1>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1309284">Mozilla Bug 1309284</a>

<script class="testbody" type="text/javascript">
"use strict";

// Execute the full-scope test
SimpleTest.waitForExplicitFinish();

SpecialPowers.pushPrefEnv({"set": [["security.webauth.webauthn", true],
                                   ["security.webauth.webauthn_enable_softtoken", true],
                                   ["security.webauth.webauthn_enable_usbtoken", false]]},
function() {
  is(navigator.authentication, undefined, "navigator.authentication does not exist any longer");
  isnot(navigator.credentials, undefined, "Credential Management API endpoint must exist");
  isnot(navigator.credentials.create, undefined, "CredentialManagement create API endpoint must exist");
  isnot(navigator.credentials.get, undefined, "CredentialManagement get API endpoint must exist");

  let credm = navigator.credentials;

  let gCredentialChallenge = new Uint8Array(16);
  window.crypto.getRandomValues(gCredentialChallenge);
  let gAssertionChallenge = new Uint8Array(16);
  window.crypto.getRandomValues(gAssertionChallenge);

  testMakeCredential();

  function decodeCreatedCredential(aCredInfo) {
    /* PublicKeyCredential : Credential
       - rawId: Key Handle buffer pulled from U2F Register() Response
       - response : AuthenticatorAttestationResponse : AuthenticatorResponse
         - attestationObject: CBOR object
         - clientDataJSON: serialized JSON
       - clientExtensionResults: (not yet supported)
    */

    ok(aCredInfo.rawId.length > 0, "Key ID exists");

    let clientData = JSON.parse(buffer2string(aCredInfo.response.clientDataJSON));
    is(clientData.challenge, bytesToBase64UrlSafe(gCredentialChallenge), "Challenge is correct");
    is(clientData.origin, window.location.origin, "Origin is correct");
    is(clientData.hashAlg, "S256", "Hash algorithm is correct");

    return webAuthnDecodeAttestation(aCredInfo.response.attestationObject.buffer)
    .then(function(decodedResult) {
      aCredInfo.clientDataObj = clientData;
      aCredInfo.publicKeyHandle = decodedResult.publicKeyHandle;
      aCredInfo.attestationObject = decodedResult.attestationObject;
      return aCredInfo;
    });
  }

  function checkAssertionAndSigValid(aPublicKey, aAssertion) {
    /* PublicKeyCredential : Credential
       - rawId: ID of Credential from AllowList that succeeded
       - response : AuthenticatorAssertionResponse : AuthenticatorResponse
         - clientDataJSON: serialized JSON
         - authenticatorData: RP ID Hash || U2F Sign() Response
         - signature: U2F Sign() Response
    */

    ok(aAssertion.rawId.length > 0, "Key ID exists");

    ok(aAssertion.response.authenticatorData.length > 0, "Authenticator data exists");
    let clientData = JSON.parse(buffer2string(aAssertion.response.clientDataJSON));
    is(clientData.challenge, bytesToBase64UrlSafe(gAssertionChallenge), "Challenge is correct");
    is(clientData.origin, window.location.origin, "Origin is correct");
    is(clientData.hashAlg, "S256", "Hash algorithm is correct");

    // Parse the signature data
    if (aAssertion.response.signature[0] != 0x01) {
      throw "User presence byte not set";
    }
    let presenceAndCounter = aAssertion.response.signature.slice(0,5);
    let signatureValue = aAssertion.response.signature.slice(5);

    let rpIdHash = aAssertion.response.authenticatorData.slice(0,32);

    // Assemble the signed data and verify the signature
    return deriveAppAndChallengeParam(window.location.host, aAssertion.response.clientDataJSON)
    .then(function(aParams) {
      console.log(aParams.appParam, rpIdHash, presenceAndCounter, aParams.challengeParam);
      console.log("ClientData buffer: ", hexEncode(aAssertion.response.clientDataJSON));
      console.log("ClientDataHash: ", hexEncode(aParams.challengeParam));
      return assembleSignedData(aParams.appParam, presenceAndCounter, aParams.challengeParam);
    })
    .then(function(aSignedData) {
      console.log(aPublicKey, aSignedData, signatureValue);
      return verifySignature(aPublicKey, aSignedData, signatureValue);
    })
  }

  function testMakeCredential() {
    let rp = {id: document.domain, name: "none", icon: "none"};
    let user = {id: "none", name: "none", icon: "none", displayName: "none"};
    let param = {type: "public-key", algorithm: "ES256"};
    let makeCredentialOptions = {
      rp: rp,
      user: user,
      challenge: gCredentialChallenge,
      parameters: [param]
    };
    credm.create({publicKey: makeCredentialOptions})
    .then(decodeCreatedCredential)
    .then(testMakeDuplicate)
    .catch(function(aReason) {
      ok(false, aReason);
      SimpleTest.finish();
    });
  }

  function testMakeDuplicate(aCredInfo) {
    let rp = {id: document.domain, name: "none", icon: "none"};
    let user = {id: "none", name: "none", icon: "none", displayName: "none"};
    let param = {type: "public-key", algorithm: "ES256"};
    let makeCredentialOptions = {
      rp: rp,
      user: user,
      challenge: gCredentialChallenge,
      parameters: [param],
      excludeList: [{type: "public-key", id: Uint8Array.from(aCredInfo.rawId),
                     transports: ["usb"]}]
    };
    credm.create({publicKey: makeCredentialOptions})
    .then(function() {
      // We should have errored here!
      ok(false, "The excludeList didn't stop a duplicate being created!");
      SimpleTest.finish();
    })
    .catch(function(aReason) {
      ok(aReason.toString().startsWith("NotAllowedError"), "Expect NotAllowedError, got " + aReason);
      testAssertion(aCredInfo);
    });
  }

  function testAssertion(aCredInfo) {
    let newCredential = {
      type: "public-key",
      id: Uint8Array.from(aCredInfo.rawId),
      transports: ["usb"],
    }

    let publicKeyCredentialRequestOptions = {
      challenge: gAssertionChallenge,
      timeout: 5000, // the minimum timeout is actually 15 seconds
      rpId: document.domain,
      allowList: [newCredential]
    };
    credm.get({publicKey: publicKeyCredentialRequestOptions})
    .then(function(aAssertion) {
      /* Pass along the pubKey. */
      return checkAssertionAndSigValid(aCredInfo.publicKeyHandle, aAssertion);
    })
    .then(function(aSigVerifyResult) {
      ok(aSigVerifyResult, "Signing signature verified");
      SimpleTest.finish();
    })
    .catch(function(reason) {
      ok(false, "Signing signature invalid: " + reason);
      SimpleTest.finish();
    });
  }
});

</script>

</body>
</html>
